Generalize the reader classes
