+++ elyra/metadata/schemas/url-catalog.json
@@ -56,7 +56,7 @@
             "category": "Component Categories"
           }
         },
-        "urls": {
+        "paths": {
           "title": "URLs",
           "description": "A list of URLs to individual component specification files",
           "type": "array",
@@ -70,7 +70,7 @@
           }
         }
       },
-      "required": ["runtime", "urls"]
+      "required": ["runtime", "paths"]
     }
   },
   "required": ["schema_name", "display_name", "metadata"]
