+++ elyra/templates/airflow/airflow_template.jinja2
@@ -44,7 +44,7 @@ from airflow.operators.{{operation.module_name}} import {{ operation.class_name 
 from airflow.contrib.operators.kubernetes_pod_operator import KubernetesPodOperator
 {% endif %}
 
-# Operator source: {{ operation.operator_source }}{% if not operation.is_generic_operator %}
+{% if operation.operator_source %}# Operator source: {{ operation.operator_source }}{% if not operation.is_generic_operator %}{% endif %}
 op_{{ operation.id|regex_replace }} = {{ operation.class_name }}(
                                                            task_id='{{ operation.notebook|regex_replace }}',
         {% for param, value in operation.component_params.items() %}
