+++ etc/config/metadata/component-registries/elyra-airflow-url-preconfig.json
@@ -4,7 +4,7 @@
     "description": "Preloaded URL-based components that are supported by Apache Airflow",
     "runtime": "airflow",
     "categories": ["Preloaded Airflow"],
-    "urls": [
+    "paths": [
       "https://raw.githubusercontent.com/apache/airflow/1.10.15/airflow/operators/bash_operator.py",
       "https://raw.githubusercontent.com/apache/airflow/1.10.15/airflow/operators/email_operator.py",
       "https://raw.githubusercontent.com/apache/airflow/1.10.15/airflow/operators/http_operator.py",
